+++ apps/frontend/src/pages/auth/register.tsx
@@ -1,9 +1,18 @@
 import { ROUTES } from "@/lib/constants";
+import { useEnabledCoreFeatures } from "@/lib/hooks/graphql";
 import { gqlClient } from "@/lib/services/api";
-import { Anchor, Box, Button, PasswordInput, TextInput } from "@mantine/core";
+import {
+	Anchor,
+	Box,
+	Button,
+	PasswordInput,
+	TextInput,
+	Tooltip,
+} from "@mantine/core";
 import { useForm, zodResolver } from "@mantine/form";
 import { notifications } from "@mantine/notifications";
 import {
+	RegisterErrorVariant,
 	RegisterUserDocument,
 	type UserInput,
 } from "@ryot/generated/graphql/backend/graphql";
@@ -11,6 +20,7 @@ import { useMutation } from "@tanstack/react-query";
 import Head from "next/head";
 import Link from "next/link";
 import { useRouter } from "next/router";
+import { match } from "ts-pattern";
 import { z } from "zod";
 
 const formSchema = z
@@ -29,6 +39,7 @@ type FormSchema = z.infer<typeof formSchema>;
 
 export default function Page() {
 	const router = useRouter();
+	const enabledFeatures = useEnabledCoreFeatures();
 	const registerUser = useMutation({
 		mutationFn: async (input: UserInput) => {
 			const { registerUser } = await gqlClient.request(RegisterUserDocument, {
@@ -38,9 +49,16 @@ export default function Page() {
 		},
 		onSuccess(data) {
 			if (data.__typename === "RegisterError") {
+				const message = match(data.error)
+					.with(RegisterErrorVariant.Disabled, () => "Registration is disabled")
+					.with(
+						RegisterErrorVariant.UsernameAlreadyExists,
+						() => "This username already exists",
+					)
+					.exhaustive();
 				notifications.show({
 					title: "Error with registration",
-					message: "This username already exists",
+					message,
 					color: "red",
 				});
 			} else {
@@ -97,9 +115,23 @@ export default function Page() {
 					{...form.getInputProps("confirm")}
 					required
 				/>
-				<Button mt="md" type="submit" loading={registerUser.isLoading} w="100%">
-					Register
-				</Button>
+				<Tooltip
+					label="Sign ups are disabled on this instance"
+					disabled={enabledFeatures.data?.signupAllowed}
+				>
+					<Button
+						mt="md"
+						type="submit"
+						{...(!enabledFeatures.data?.signupAllowed
+							? { "data-disabled": true }
+							: {})}
+						loading={registerUser.isLoading}
+						sx={{ "&[data-disabled]": { pointerEvents: "all" } }}
+						w="100%"
+					>
+						Register
+					</Button>
+				</Tooltip>
 				<Box mt="lg" style={{ textAlign: "right" }}>
 					Already a member? Login{" "}
 					<Link href={ROUTES.auth.login} passHref legacyBehavior>